requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2020.6.20
Django==4.0.3
djangorestframework==3.13.1
flake8==4.0.1
mccabe==0.6.1
Pillow==9.0.1
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2021.3
sqlparse==0.4.2
